refactor event sorting
